+++ source/ProcessManager.Example.Orchestrations/Processes/BRS_X02/ActorRequestProcessExample/V1/Orchestration/Steps/EnqueueActorMessagesStep.cs
@@ -35,7 +35,7 @@ internal class EnqueueActorMessagesStep(
 
     protected override int StepSequenceNumber => StepSequence;
 
-    protected override async Task<OrchestrationStepTerminationState> OnExecuteAsync()
+    protected override async Task<StepInstanceTerminationState> OnExecuteAsync()
     {
         var enqueueIdempotencyKey = Context.NewGuid();
 
@@ -67,6 +67,6 @@ internal class EnqueueActorMessagesStep(
             eventName: ActorRequestProcessExampleNotifyEventV1.OrchestrationInstanceEventName,
             timeout: waitTimeout);
 
-        return OrchestrationStepTerminationState.Succeeded;
+        return StepInstanceTerminationState.Succeeded;
     }
 }